refactor: retire pipeline_builder_tab, merge into DesignModeCanvas; add wf-* CSS
- Renamed Build tab to Workflow in CanvasBackFace; now Sources + Workflow only
  (History was already superseded by MAP mode in CanvasWorkflowPipeline)
- Ported skeleton-block pre-creation and run_pipeline/run_saved_pipeline logic
  from pipeline_builder_tab into DesignModeCanvas run_workflow handler
- Deleted apps/papillon/frontend/src/components/pipeline_builder_tab.rs
- Removed pub mod pipeline_builder_tab from components/mod.rs
- Appended all wf-* CSS classes to main.css: node anatomy, port rows with
  connection dots (input left-edge / output right-edge), edge states (teal solid,
  gold dashed, coral blocked), pulse animation, tools strip, template picker,
  TTL badge, and inline EdgeApprovalCard with Allow/AlwaysAllow/Deny actions
- cargo check: zero errors; cargo test -p papillon-shared: 307 passed
- cargo test -p papillon port_compat: 5 passed
