+++ apps/papillon/frontend/src/components/mod.rs
@@ -6,7 +6,6 @@ pub mod canvas_empty_state;
 pub mod canvas_workflow_pipeline;
 pub mod hitl_gate;
 pub mod outcome_summary;
-pub mod pipeline_builder_tab;
 pub mod profile_avatar;
 pub mod recovery_setup;
 pub mod registry;
